phase2 part3 done

+++ app/Http/Controllers/Admin/AdminController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CreateAdminValidation;
+use App\Http\Requests\UpdateAdminValidation;
 use App\Repositories\AdminRepositoryInterface;
 use Illuminate\Http\Request;
 
@@ -20,6 +22,7 @@ class AdminController extends Controller
 
     public function index()
     {
+        $this->authorize('admin-viewAll-admins');
         $admins = $this->adminRepository->all();
         return view('admin.admins.index')->with('admins', $admins);
     }
@@ -27,11 +30,12 @@ class AdminController extends Controller
 
     public function create()
     {
+        $this->authorize('admin-create-admin');
         return view('admin.admins.create');
     }
 
 
-    public function store(Request $request)
+    public function store(CreateAdminValidation $request)
     {
         $this->adminRepository->save($request);
         return redirect()->route('admins.index');
@@ -40,12 +44,13 @@ class AdminController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('admin-edit-admin');
         $admin = $this->adminRepository->find($id);
         return view('admin.admins.edit')->with('admin', $admin);
     }
 
 
-    public function update(Request $request, $id)
+    public function update(UpdateAdminValidation $request, $id)
     {
         $this->adminRepository->update($request, $id);
         return redirect()->route('admins.index');
@@ -54,6 +59,7 @@ class AdminController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('admin-delete-admin');
         $this->adminRepository->destroy($id);
         return redirect()->route('admins.index');
     }
